requirement.txt
Flask==2.2.5
gunicorn==20.1.0
google-generativeai==0.7.2
python-dotenv==1.0.1
